Use static libraries (release review)

+++ 3dEngine/src/graphic/shader/data/ShaderDataSender.cpp
@@ -13,7 +13,7 @@ namespace urchin {
 
     ShaderDataSender& ShaderDataSender::sendData(const ShaderVar& shaderVar, unsigned int count, const float* values) {
         shaderVar.getShader()->bind();
-        glUniform1fv(shaderVar.getVariableLocation(), count, values);
+        glUniform1fv(shaderVar.getVariableLocation(), (int)count, values);
         return *this;
     }
 
@@ -25,7 +25,7 @@ namespace urchin {
 
     ShaderDataSender& ShaderDataSender::sendData(const ShaderVar& shaderVar, unsigned int count, const int* values) {
         shaderVar.getShader()->bind();
-        glUniform1iv(shaderVar.getVariableLocation(), count, values);
+        glUniform1iv(shaderVar.getVariableLocation(), (int)count, values);
         return *this;
     }
 
@@ -37,7 +37,7 @@ namespace urchin {
 
     ShaderDataSender& ShaderDataSender::sendData(const ShaderVar& shaderVar, unsigned int count, const unsigned int* values) {
         shaderVar.getShader()->bind();
-        glUniform1uiv(shaderVar.getVariableLocation(), count, values);
+        glUniform1uiv(shaderVar.getVariableLocation(), (int)count, values);
         return *this;
     }
 
@@ -53,7 +53,7 @@ namespace urchin {
 
     ShaderDataSender& ShaderDataSender::sendData(const ShaderVar& shaderVar, unsigned int count, const Matrix2<float>* values) {
         shaderVar.getShader()->bind();
-        glUniformMatrix2fv(shaderVar.getVariableLocation(), count, GL_FALSE, (const float*)values);
+        glUniformMatrix2fv(shaderVar.getVariableLocation(), (int)count, GL_FALSE, (const float*)values);
         return *this;
     }
 
@@ -65,7 +65,7 @@ namespace urchin {
 
     ShaderDataSender& ShaderDataSender::sendData(const ShaderVar& shaderVar, unsigned int count, const Matrix3<float>* values) {
         shaderVar.getShader()->bind();
-        glUniformMatrix3fv(shaderVar.getVariableLocation(), count, GL_FALSE, (const float*)values);
+        glUniformMatrix3fv(shaderVar.getVariableLocation(), (int)count, GL_FALSE, (const float*)values);
         return *this;
     }
 
@@ -77,7 +77,7 @@ namespace urchin {
 
     ShaderDataSender& ShaderDataSender::sendData(const ShaderVar& shaderVar, unsigned int count, const Matrix4<float>* values) {
         shaderVar.getShader()->bind();
-        glUniformMatrix4fv(shaderVar.getVariableLocation(), count, GL_FALSE, (const float*)values);
+        glUniformMatrix4fv(shaderVar.getVariableLocation(), (int)count, GL_FALSE, (const float*)values);
         return *this;
     }
 
@@ -113,7 +113,7 @@ namespace urchin {
 
     ShaderDataSender& ShaderDataSender::sendData(const ShaderVar& shaderVar, unsigned int count, const Vector2<float>* values) {
         shaderVar.getShader()->bind();
-        glUniform2fv(shaderVar.getVariableLocation(), count, (const float*)values);
+        glUniform2fv(shaderVar.getVariableLocation(), (int)count, (const float*)values);
         return *this;
     }
 
@@ -125,7 +125,7 @@ namespace urchin {
 
     ShaderDataSender& ShaderDataSender::sendData(const ShaderVar& shaderVar, unsigned int count, const Vector3<float>* values) {
         shaderVar.getShader()->bind();
-        glUniform3fv(shaderVar.getVariableLocation(), count, (const float*)values);
+        glUniform3fv(shaderVar.getVariableLocation(), (int)count, (const float*)values);
         return *this;
     }
 
@@ -137,7 +137,7 @@ namespace urchin {
 
     ShaderDataSender& ShaderDataSender::sendData(const ShaderVar& shaderVar, unsigned int count, const Vector4<float>* values) {
         shaderVar.getShader()->bind();
-        glUniform4fv(shaderVar.getVariableLocation(), count, (const float*)values);
+        glUniform4fv(shaderVar.getVariableLocation(), (int)count, (const float*)values);
         return *this;
     }
